requirements.txt
absl_py==1.1.0
chex==0.1.3
clu==0.0.7
datasets==2.3.2
distrax==0.1.2
flax==0.5.2
h5py==3.7.0
ipykernel==6.15.1
matplotlib==3.5.2
ml_collections==0.1.1
numpy==1.23.1
opt_einsum==3.3.0
optax==0.1.3
setuptools==59.5.0
tensorflow_datasets==4.6.0
torchvision==0.13.0
tqdm==4.64.0
wandb==0.12.21
